Adding vsync option. Pulling out the X264_PROFILE enum into its own enum class. Doesn’t seem to fit our naming pattern?
